requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.0.0
streamlit>=1.0.0
pillow>=9.0.0
requests>=2.28.0
beautifulsoup4>=4.9.0 